LPS-102159 Fix split package issue
