+++ portal-impl/test/unit/com/liferay/portal/dao/db/PostgresDBTest.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.dao.db;
 
-import com.liferay.portal.kernel.dao.db.BaseDBTestCase;
 import com.liferay.portal.kernel.dao.db.DB;
+import com.liferay.portal.kernel.dao.db.test.BaseDBTestCase;
 
 import java.io.IOException;
