+++ portal-test/bnd.bnd
@@ -2,7 +2,7 @@ Bundle-Name: ${manifest.bundle.name}
 Bundle-SymbolicName: ${manifest.bundle.symbolic.name}
 Bundle-Version: 7.0.0
 Export-Package:\
-	com.liferay.portal.kernel.dao.db,\
+	com.liferay.portal.kernel.dao.db.test,\
 	com.liferay.portal.kernel.test,\
 	com.liferay.portal.kernel.test.context,\
 	com.liferay.portal.kernel.test.junit,\
