+++ portal-test/src/com/liferay/portal/kernel/dao/db/test/BaseDBTestCase.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.kernel.dao.db;
+package com.liferay.portal.kernel.dao.db.test;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.dao.db.DB;
 
 import java.io.IOException;
